fix: 🐛 Resolve errors in Storybook
