+++ src/components/Radio/Radio.types.ts
@@ -1,0 +1,5 @@
+import { RadioProps as MUIRadioProps } from '@material-ui/core/Radio';
+
+
+export type RadioProps = MUIRadioProps;
+
