+++ src/components/Radio/index.tsx
@@ -1,32 +1,2 @@
-
-
-import * as React from 'react';
-import MuiRadio from '@material-ui/core/Radio';
-import { Props } from './types';
-
-
-/**
- * Radio buttons allow the user to select one option from a set.
- *
- * [Material-UI Docs](https://material-ui.com/components/radio-buttons/)
- *
- *
- * > No differences in Brewkit.
- *
- */
-export const Radio = React.forwardRef(({
-    ...otherProps
-}: Props, ref: React.Ref<any>): React.ReactElement => (
-
-
-    <MuiRadio ref={ref} {...otherProps} />
-
-
-));
-
-
-Radio.displayName = 'Radio';
-
-
-export default Radio;
-
+export { default } from './Radio';
+export type { RadioProps } from './Radio.types';
