--- src/components/Radio/types.ts
@@ -1,5 +1,0 @@
-import { RadioProps } from '@material-ui/core/Radio';
-
-
-export type Props = RadioProps;
-
